refactor(ParticipantTest): Participant 메소드에 관한 테스트 PlayerTest -> ParticipantTest로 이동
